chore: rename rollback file to compaction
